requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.1
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.31
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
